Extract Base OAuth Class
